chore: refactor hyperdeck into multiple files
